cxj commit 2018/07/27 2:08

+++ src/com/ssm/model/bean/AuctionOrder.java
@@ -15,7 +15,7 @@ public class AuctionOrder {
 
     private String address;
     private int postcode;
-    private int acceptPhone;
+    private long acceptPhone;
     private String acceptName;
 
     public String getAddress() {
@@ -34,11 +34,11 @@ public class AuctionOrder {
         this.postcode = postcode;
     }
 
-    public int getAcceptPhone() {
+    public long getAcceptPhone() {
         return acceptPhone;
     }
 
-    public void setAcceptPhone(int acceptPhone) {
+    public void setAcceptPhone(long acceptPhone) {
         this.acceptPhone = acceptPhone;
     }
